python-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
pydantic==2.5.0
motor==3.3.2